Broke down tests and nit fixes
